oups mea culpa !!!!
je ne me suis pas relu, j'avais déjà une vilaine fonction define_color qui trainait mais uniquement centrée sur le CFT mais là ca rentre dans l'ordre........
